# Better capacity and growth rate for `{Stream,Future}Group`