--- a/src/collections/mod.rs
+++ b/src/collections/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod group_inner;
 #[cfg(feature = "alloc")]
 pub mod vec;