--- a/src/collections/mod.rs
+++ b/src/collections/mod.rs
@@ -1,3 +1,3 @@
-pub(crate) mod group_inner;
+pub(crate) mod inner_group;
 #[cfg(feature = "alloc")]
 pub mod vec;