--- a/src/collections/mod.rs
+++ b/src/collections/mod.rs
@@ -1,3 +1,4 @@
+#[cfg(feature = "alloc")]
 pub(crate) mod inner_group;
 #[cfg(feature = "alloc")]
 pub mod vec;