Fixes the same problem as #168 by keeping up with the capacity and growth rate of the group.

When a group tries to grow over its capacity, we double the previous capacity.

`FutureGroup` and `StreamGroup` are structurally very similar, they only differ on the `Stream` implementation, so I isolated the common structure to reduce the duplicated code.

## Performance

Compared to the main branch, we gained 99,775% using [this benchmark](https://github.com/notgull/futures-concurrency-benchmark) (with 1_000_000 futures).

Before:
```rust
group/futures_concurrency::FutureGroup
                        time:   [27.222 s 28.561 s 29.720 s]
```


After:

```rust
group/futures_concurrency::FutureGroup
                        time:   [68.507 ms 68.598 ms 68.752 ms]
                        change: [-99.769% -99.760% -99.748%] (p = 0.00 < 0.05)
                        Performance has improved.
```

## todos
- [x] make `FutureGroup` and `StreamGroup` type aliases for a common struct generic over the poll behavior
- [x] rename `resize` to `reserve` and expose it
- [x] tests and docs and benchmarks?

Closes #169