@matheus-consoli 2024-04-08
sure, i agree with you

my initial idea was to write a single type `InnerGroup` and have `{Future, Stream}Group` be type aliases (that would, more or less, break the funnel-shaped problem), but i also didn't want to expose those in the public API... and i wasn't considering an `AsyncIteratorGroup`.

I will close this PR and open a simpler one.